requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
alsa-lib=1.2.13=hb9d3cd8_0
appdirs=1.4.4=pyhd3eb1b0_0
beautifulsoup4=4.12.3=py312h06a4308_0
blas=1.0=mkl
bottleneck=1.4.2=py312ha883a20_0
brotli-python=1.0.9=py312h6a678d5_9
bs4=4.12.3=hd8ed1ab_1
bzip2=1.0.8=h5eee18b_6
c-ares=1.34.4=hb9d3cd8_0
ca-certificates=2024.12.31=h06a4308_0
cairo=1.18.2=h3394656_1
certifi=2024.12.14=py312h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.3.1=py312hdb19cb5_0
cuda-cudart=12.4.127=h99ab3db_0
cuda-cudart_linux-64=12.4.127=hd681fbe_0
cuda-cupti=12.4.127=h6a678d5_1
cuda-libraries=12.4.1=h06a4308_1
cuda-nvrtc=12.4.127=h99ab3db_1
cuda-nvtx=12.4.127=h6a678d5_1
cuda-opencl=12.4.127=h6a678d5_0
cuda-runtime=12.4.1=hb982923_0
cuda-version=12.4=hbda6634_3
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
double-conversion=3.3.0=h59595ed_0
expat=2.6.4=h6a678d5_0
filelock=3.13.1=py312h06a4308_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.15.0=h7e30c49_1
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.55.3=py312h5eee18b_0
freetype=2.12.1=h267a509_2
frozendict=2.4.2=py312h06a4308_0
glib=2.82.2=h07242d1_1
glib-tools=2.82.2=h4833e2c_1
graphite2=1.3.14=h295c915_1
greenlet=3.1.1=py312h6a678d5_0
harfbuzz=9.0.0=hda332d3_1
html5lib=1.1=pyhd3eb1b0_0
icu=75.1=he02047a_0
idna=3.7=py312h06a4308_0
intel-openmp=2022.0.1=h06a4308_3633
jinja2=3.1.5=py312h06a4308_0
joblib=1.4.2=py312h06a4308_0
kiwisolver=1.4.8=py312h6a678d5_0
krb5=1.21.3=h143b758_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20240722.0=cxx17_hbbce691_4
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libclang-cpp18.1=18.1.8=default_hf981a13_5
libclang13=19.1.7=default_h9c6a7e4_0
libcublas=12.4.5.8=h99ab3db_1
libcufft=11.2.1.3=h99ab3db_1
libcufile=1.9.1.3=h99ab3db_1
libcups=2.3.3=h4637d8d_4
libcurand=10.3.5.147=h99ab3db_1
libcurl=8.11.1=h332b0f4_0
libcusolver=11.6.1.9=h99ab3db_1
libcusparse=12.3.1.170=h99ab3db_1
libdeflate=1.23=h4ddbbb0_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20230828=h5eee18b_0
libegl=1.7.0=ha4b6fd6_2
libev=4.33=h7f8727e_1
libexpat=2.6.4=h5888daf_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_2
libglib=2.82.2=h2ff4ddf_1
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h77fa898_1
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.3=h5eee18b_0
liblapack=3.9.0=16_linux64_mkl
libllvm18=18.1.8=h8b73ec9_2
libllvm19=19.1.7=ha7bfdaf_0
liblzma=5.6.3=hb9d3cd8_1
libnghttp2=1.64.0=h161d5f1_0
libnpp=12.2.5.30=h99ab3db_1
libnsl=2.0.1=hd590300_0
libnvfatbin=12.4.127=h7934f7d_2
libnvjitlink=12.4.127=h99ab3db_1
libnvjpeg=12.3.1.117=h6a678d5_1
libopenblas=0.3.21=h043d6bf_0
libopengl=1.7.0=ha4b6fd6_2
libpciaccess=0.18=hd590300_0
libpng=1.6.45=h943b412_0
libpq=16.6=h035377e_1
libprotobuf=5.27.5=h5b01275_2
libsqlite=3.48.0=hee588c1_0
libssh2=1.11.1=hf672d98_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtiff=4.7.0=hd9ff511_3
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.5=h8d12d68_1
libxslt=1.1.41=h097e994_0
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=15.0.7=h0cdce71_0
lxml=5.3.0=py312h57af460_1
lz4-c=1.9.4=h6a678d5_1
mariadb=1.1.11=py312h2ec8cdc_0
mariadb-connector-c=3.4.3=h10329db_1
markupsafe=2.1.3=py312h5eee18b_1
matplotlib=3.10.0=py312h7900ff3_0
matplotlib-base=3.10.0=py312hbfdbfaf_0
mkl=2022.1.0=hc2b9512_224
mpmath=1.3.0=py312h06a4308_0
multitasking=0.0.9=pyhd8ed1ab_0
mysql=9.0.1=h27aab58_1
mysql-client=9.0.1=he0572af_1
mysql-common=9.0.1=h266115a_1
mysql-devel=9.0.1=h266115a_1
mysql-libs=9.0.1=he0572af_1
mysql-server=9.0.1=h06f4e53_1
ncurses=6.4=h6a678d5_0
networkx=3.4.2=py312h06a4308_0
nomkl=2.0=0
numexpr=2.10.2=py312h6a710ac_100
numpy=2.2.2=py312h72c5963_0
ocl-icd=2.3.2=h5eee18b_1
openjpeg=2.5.3=h5fbd93e_0
openssl=3.4.0=h7b32b05_1
packaging=24.2=py312h06a4308_0
pandas=2.2.3=py312h6a678d5_0
pcre2=10.44=hba22ea6_2
peewee=3.17.7=py312ha6a020b_0
pillow=11.1.0=py312h80c1187_0
pip=24.2=py312h06a4308_0
pixman=0.44.2=h29eaf8c_0
platformdirs=3.10.0=py312h06a4308_0
pthread-stubs=0.3=h0ce48e5_1
pymysql=1.1.1=pyhd8ed1ab_1
pyparsing=3.2.0=py312h06a4308_0
pyside6=6.7.2=py312h91f0f75_4
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0post0=py312h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=5_cp312
pytorch=2.5.1=py3.12_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2024.1=py312h06a4308_0
pyyaml=6.0.2=py312h5eee18b_0
qt6-main=6.7.2=hb12f9c5_5
readline=8.2=h5eee18b_0
requests=2.32.3=py312h06a4308_1
scikit-learn=1.5.2=py312h6a678d5_0
scipy=1.15.1=py312h180e4f1_0
setuptools=75.1.0=py312h06a4308_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.5=py312h06a4308_0
sqlalchemy=2.0.37=py312h66e93f0_0
sqlite=3.31.1=h7b6447c_0
sympy=1.13.3=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py312he106c6f_0
tk=8.6.13=noxft_h4845f30_101
torchtriton=3.1.0=py312
tornado=6.4.2=py312h5eee18b_0
typing-extensions=4.12.2=py312h06a4308_0
typing_extensions=4.12.2=py312h06a4308_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=py312h5eee18b_1
urllib3=2.3.0=py312h06a4308_0
wayland=1.23.1=h3e06ad9_0
webencodings=0.5.1=py312h06a4308_2
wheel=0.44.0=py312h06a4308_0
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.4=h5eee18b_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.5=he73a12e_0
xorg-libx11=1.8.10=h4f16b4b_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7f98852_2
yfinance=0.2.52=pyhd8ed1ab_0
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.6=ha6fb4c9_0
